+++ semitemplates/ts/.eslintrc.js
@@ -1,14 +1,10 @@
-export function get_eslintrc(): string {
-  // https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/README.md
-  // could add jest?
-  return (
-    `module.exports = {
+module.exports = {
   "env": {
     "es2021": true,
     "node": true
   },
   "extends": [
-    "@srbrahma/eslint-config",
+    "eslint-config-gev",
     "plugin:@typescript-eslint/recommended-requiring-type-checking" // https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/TYPED_LINTING.md#getting-started---linting-with-type-information
   ],
   "parser": "@typescript-eslint/parser",
@@ -22,5 +18,3 @@ export function get_eslintrc(): string {
   }
 };
 
-`);
-}