+++ src/resources/get_index.ts
@@ -1,5 +1,0 @@
-export function get_index(): string {
-  return (
-    `console.log('Howdy World!');`
-  );
-}